Fix main route prerender: client-only layout shell
Split (main) layout into server wrapper + MainLayoutClient.
Use dynamic(ssr: false) and force-dynamic to match prior SPA
behavior and avoid window/localStorage access during static generation.
